fix[install_standard.sh]: Fix ytop install

+++ .bin/install_standard.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 sudo add-apt-repository ppa:mmstick76/alacritty -n -y
-sudo add-apt-repository ppa:regolith-linux/release -n -y
+sudo add-apt-repository ppa:kgilmer/speed-ricer -n -y
 sudo add-apt-repository ppa:lazygit-team/release -n -y
 curl -sL https://deb.nodesource.com/setup_12.x | sudo -E bash -
 sudo apt-get update
@@ -60,14 +60,16 @@ done
 fc-cache -fv
 cd ~
 
-# Ytop
-ytop_path="/tmp/ytop.tar.gz"
-wget https://github.com/cjbassi/ytop/releases/download/0.6.2/ytop-0.6.2-x86_64-unknown-linux-gnu.tar.gz -O "$ytop_path"
-tar xzf -C "$HOME/.bin" "$ytop_path"
-rm -fr "$ytop_path"
+if ! command -v ytop > /dev/null; then
+	ytop_path="/tmp/ytop.tar.gz"
+	wget "https://github.com/cjbassi/ytop/releases/download/0.6.2/ytop-0.6.2-x86_64-unknown-linux-gnu.tar.gz" -O "$ytop_path" &&
+	tar -xzf "$ytop_path" -C "$HOME/.bin"
+	rm -fr "$ytop_path"
+fi
 
-# lsd
-lsd_path="/tmp/lsd_18.deb"
-wget https://github.com/Peltoche/lsd/releases/download/0.18.0/lsd_0.18.0_amd64.deb -O "$lsd_path"
-sudo dpkg -i "$lsd_path"
-rm -fr "$lsd_path"
+if ! command -v "lsd" > /dev/null; then
+	lsd_path="/tmp/lsd_18.deb"
+	wget "https://github.com/Peltoche/lsd/releases/download/0.18.0/lsd_0.18.0_amd64.deb" -O "$lsd_path" &&
+	sudo dpkg -i "$lsd_path"
+	rm -fr "$lsd_path"
+fi
